Fix discrepancies btw logged in user & submission
In some apps, for example app-worship-decisions-database, the logged in
user can be different from the bestuur that created the submission we're
trying to get info from.

Ex: a Gemeente can be opening a submission sent by a CKB.

In this case, the check `isCKB`, based on the logged in user only, is
failing to help the service follow the correct path. I added an extra
check on the type of bestuuren that are allowed to decide on the type of
submissions that we request extra information on.