Remove always false function
The function 'isDecisionTypeFromCKB' always evaluates to 'false', given
the configuration. We get correct results, thus we can remove this
function and its calls. Checking if the decision types is from a CKB is
also indirectly checked by the following lines. Line
https://github.com/lblod/worship-decisions-cross-reference-service/blob/a5a61e72e5d4f82d7675da629d6dbdac3af17de1/app.js#L122
will evaluate to 'undefined' if it cannot find the type and so we end up
in the same situation as if the type was not sent in by the CKB in the
first place.