Some improvements (#8)
* Remove hardcoded URI from testing(?)

Why is this URI there? This refers to some random bestuurseenheid in
Belgium. This was probably used for testing?

* Remove always false function

The function 'isDecisionTypeFromCKB' always evaluates to 'false', given
the configuration. We get correct results, thus we can remove this
function and its calls. Checking if the decision types is from a CKB is
also indirectly checked by the following lines. Line
https://github.com/lblod/worship-decisions-cross-reference-service/blob/a5a61e72e5d4f82d7675da629d6dbdac3af17de1/app.js#L122
will evaluate to 'undefined' if it cannot find the type and so we end up
in the same situation as if the type was not sent in by the CKB in the
first place.

* Revert "Remove always false function"

This reverts commit b41f7fb5799bbd055b0b4e1cd8120f2b426f6555.